04_building_effective_agents/04_augumentation_memory/06_graphiti_learning_path/11_teaching_agent_capstone/tutorsgpt_mcp_server/pyproject.toml
[project]
name = "tutorsgpt-mcp-server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.11.1",
    "graphiti-core>=0.18.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
